Add regular reachability algorithm
This commit adds an implementation of the regular reachability algorithm based
on linear-algebra graph processing approach. The algorithm finds a set of nodes
in a directed-labelled graph. These nodes are reachable by paths starting from
one of source nodes and having edges labels conforming a word satisfying
restrictions defined by a regular language.

This algorithm is based on the bread-first-search algorithm over the adjacency
matrices. Regular restrictions are definied by non-deterministic finite
automata. The algorithm considers the paths on which "label words" are accepted
by the specified NFA.

Algorithm is used with the following inputs:
- A regular automaton adjacency matrix decomposition.
- A graph adjacency matrix decomposition.
- An array of the starting node indicies.

It results with a vector, having v[i] = 1 iff the node is reachable by a
path satisfying the regular restrictions.

More information is available at http://arxiv/will/be/here