Fix regular path query test coverage
In #261 an algorithm for evaluating regular path queries has been
introduced. The algorithm uses NFAs for representing regular
constraints and uses transpositions of their adjacency matrices. The
implementation involved checks on whether the provided adjacency matrix
is symmetric and its transpositions are cached though this behavior
hasn't been tested and the coverage has been lower than 100%. This patch
fixes it by varying whether to cache the matrix transpositions and
compute the symmetry flags.